refactor(FormGroup): ラベルの描画をFormControl・Fieldsetに移譲する
LabelClusterはisFieldsetによってlegendとlabelを出し分けており、1つの
コンポーネントが2種類のマークアップを抱えていた。それぞれの持ち主である
FormControlとFieldsetに分離し、FormGroupはLabelComponentをpropsとして
受け取るようにした。

これによりFormGroupからisFieldsetによる分岐が1つ減り、Fieldset側は
legend固定、FormControl側はlabel固定の実装になる。

あわせてFormControlでは、subActionAreaが無い場合にCluster
（justify="space-between"）でラップしないようにした。DOM構造が
変わるため、Chromaticに差分が出る可能性がある。

生成DOMは、FormControl・Fieldsetそれぞれの基本形・statusLabels・
subActionArea・unrecommendedHide・styleType指定・可視ラベル無しinputへの
aria-label付与を含む11パターンで実測比較済み。上記のラップ削除に該当する
3ケース以外はmasterと一致している。
