+++ packages/smarthr-ui/src/components/FormGroup/FormControl.tsx
@@ -1,12 +1,15 @@
 'use client'
 
-import { type FC, type ReactNode, useEffect, useId, useRef, useState } from 'react'
+import { type FC, type ReactNode, memo, useEffect, useId, useRef, useState } from 'react'
 
 import { useObjectAttributes } from '../../hooks/useObjectAttributes'
+import { Cluster } from '../Layout'
+import { Text } from '../Text'
+import { VisuallyHiddenText } from '../VisuallyHiddenText'
 
 import { CHILDREN_WRAPPER_INPUT_SELECTOR, FormGroup } from './FormGroup'
 
-import type { CommonProps, ObjectLabelType } from './type'
+import type { CommonProps, LabelComponentProps, ObjectLabelType } from './type'
 
 const labelObjectConverter = (label: ReactNode) => ({ text: label })
 
@@ -62,5 +65,61 @@ export const FormControl: FC<
     }
   }, [label.htmlFor, label.id])
 
-  return <FormGroup {...rest} wrapperRef={wrapperRef} label={label} />
+  return (
+    <FormGroup {...rest} wrapperRef={wrapperRef} label={label} LabelComponent={LabelComponent} />
+  )
 }
+
+const LabelComponent = memo<LabelComponentProps>(
+  ({
+    managedHtmlFor,
+    managedLabelId,
+    unrecommendedHideLabel,
+    labelType = 'blockTitle',
+    label,
+    labelIcon,
+    subActionArea,
+    labelClassName,
+    statusLabels,
+  }) => {
+    const body = (
+      <>
+        <Text styleType={labelType} icon={labelIcon}>
+          <span className="smarthr-ui-FormControl-labelText">{label}</span>
+        </Text>
+        {statusLabels.length > 0 && (
+          <Cluster gap={0.25} as="span">
+            {statusLabels}
+          </Cluster>
+        )}
+      </>
+    )
+
+    const attrs = {
+      as: 'label' as const,
+      htmlFor: managedHtmlFor,
+      id: managedLabelId,
+    }
+
+    if (unrecommendedHideLabel) {
+      return <VisuallyHiddenText {...attrs}>{body}</VisuallyHiddenText>
+    }
+
+    const renderedLabel = (
+      <Cluster {...attrs} align="center" className={labelClassName}>
+        {body}
+      </Cluster>
+    )
+
+    if (subActionArea) {
+      return (
+        <Cluster justify="space-between">
+          {renderedLabel}
+          <div className="shr-grow">{subActionArea}</div>
+        </Cluster>
+      )
+    }
+
+    return renderedLabel
+  },
+)
